requirements.txt
# Smart Attendance System v8.0 — Requirements
# Install: pip install -r requirements.txt

# Core CV
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# ML
scikit-learn>=1.3.0

# Database
psycopg2-binary>=2.9.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
PyJWT>=2.8.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Environment config (FIX #5,6,7)
python-dotenv>=1.0.0

# Data
pandas>=2.0.0

# Face recognition
cmake
dlib
face-recognition

# Skeleton liveness (FIX #12: shared instance)
mediapipe>=0.10.0

# Optional: Excel/PDF export
openpyxl>=3.1.0

# Ensure python-dotenv is present (for .env loading)
python-dotenv>=1.0.0
